Address Copilot review feedback
- Extract duplicated inference-range helpers from segmentation.py and
  instance_segmentation.py into a new InferenceRangeMixin in
  qgis_plugin/geoai/dialogs/inference_range.py so the two dock widgets
  share one implementation and cannot drift apart.
- Replace tempfile.NamedTemporaryFile(delete=False) with
  tempfile.mkdtemp() + joined clip.tif path in both inference workers.
  This avoids leaving an empty file at the target path before
  clip_raster_by_bbox writes it (some GDAL drivers reject pre-existing
  outputs). The temp directory is removed with shutil.rmtree in the
  finally block.
